Refactor BlobStoreStats test classes to new file
